+++ tests/test_settings.py
@@ -2781,7 +2781,7 @@ def main() -> None:
     defaults = {}
     setting_ids = set()
 
-    assert data.version == "3.0.0"
+    assert data.version == "3.0.1"
 
     gradient_name = localization["mod_name"]["en"]
     assert gradient_name.startswith("{#color(174,239,105)}B")